+++ src/app/admin/components/admin-product-edit/admin-product-edit.component.ts
@@ -1,15 +1,9 @@
 import { Component } from '@angular/core';
 import { Product } from 'src/app/shared/interfaces/product';
-import { ActivatedRoute, Router } from '@angular/router';
-import { ProductServiceService } from 'src/app/product/services/product-service.service';
+import { ActivatedRoute } from '@angular/router';
+import { ProductService } from 'src/app/product/services/product.service';
 import { CartService } from 'src/app/cart/services/cart.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
-
-enum ProductSearchStatus {
-  Undefined,
-  Found,
-  NotFound
-}
 
 @Component({
   selector: 'app-admin-product-edit',
@@ -18,16 +12,12 @@ enum ProductSearchStatus {
 })
 export class AdminProductEditComponent {
   product: Product = {} as Product;
-  testValue: string;
   productIdUrlParam: string;
-  public productSearchStatus: ProductSearchStatus = ProductSearchStatus.Undefined;
-  productTotalRemaining: number;
-  isAllProductDataLoaded: boolean;
   isNewMode: boolean;
 
   constructor(
     private route: ActivatedRoute,
-    private productService: ProductServiceService,
+    private productService: ProductService,
     private cartService: CartService,
     private snackBar: MatSnackBar
   ) {
@@ -35,18 +25,25 @@ export class AdminProductEditComponent {
   }
 
   onSaveClick() {
-    if (this.isNewMode === false) {
-      this.productService.updateProductInStock(this.product, this.productTotalRemaining);
-    } else if (this.isNewMode === true) {
-      this.productService.addNewProductInStock(this.product, this.productTotalRemaining);
+    if (this.isNewMode) {
+      this.productService.createProduct(this.product);
+      this.isNewMode = false;
+      this.showProductCreatedPopup();
+    } else {
+      this.productService.updateProduct(this.product);
+      this.cartService.savePurchasedProducts();
+      this.showProductUpdatedPopup();
     }
-    this.cartService.savePurchasedProducts();
-    this.isNewMode = undefined;
-    this.showProductSavedPopup();
   }
 
-  showProductSavedPopup() {
-    this.snackBar.open('Product saved', 'Save product', {
+  showProductCreatedPopup() {
+    this.snackBar.open('Product created', 'Save product', {
+      duration: 2000
+    });
+  }
+
+  showProductUpdatedPopup() {
+    this.snackBar.open('Product updated', 'Save product', {
       duration: 2000
     });
   }
@@ -56,39 +53,22 @@ export class AdminProductEditComponent {
     if (this.productIdUrlParam == null) {
       this.isNewMode = true;
       this.createNewProduct();
-      this.isAllProductDataLoaded = this.productSearchStatus === ProductSearchStatus.Found;
     } else {
       this.isNewMode = false;
-      this.setProduct((p: Product) => {
-        this.setTotalRemaining(p, () => {
-          this.isAllProductDataLoaded = this.productSearchStatus === ProductSearchStatus.Found;
-        });
-      });
+      this.setProduct();
     }
   }
 
-  setProduct(callback?: (p: Product) => any) {
+  setProduct() {
     this.productService.getProduct(this.productIdUrlParam).subscribe((p) => {
       this.product = p;
-      this.productSearchStatus = p ? ProductSearchStatus.Found : ProductSearchStatus.NotFound;
-      callback?.call(this, p);
     });
 
     // TODO why this does not work?
-    // this.productService.getProduct(productId).toPromise().then(val => console.log(val));
-  }
-
-  setTotalRemaining(p: Product, callback?: () => any) {
-    if (this.productSearchStatus === ProductSearchStatus.Found) {
-      this.productService.getProductTotalRemaining(p.Id).subscribe((totalRemaining => {
-        this.productTotalRemaining = totalRemaining;
-        callback?.call(this);
-      }));
-    }
+    // this.productService.getProduct(this.productIdUrlParam).toPromise().then(val => console.log(val));
   }
 
   createNewProduct() {
     this.product = this.productService.getNewProduct();
-    this.productSearchStatus = ProductSearchStatus.Found;
   }
 }