+++ src/app/admin/components/admin-products-list/admin-products-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { ProductInStock } from 'src/app/shared/interfaces/product-in-stock';
-import { ProductServiceService } from 'src/app/product/services/product-service.service';
+import { ProductService } from 'src/app/product/services/product.service';
+import { Product } from 'src/app/shared/interfaces/product';
 
 @Component({
   selector: 'app-admin-products-list',
@@ -9,17 +9,21 @@ import { ProductServiceService } from 'src/app/product/services/product-service.
   styleUrls: ['./admin-products-list.component.scss']
 })
 export class AdminProductsListComponent implements OnInit {
-  productsInStock: Observable<ProductInStock[]>;
+  products: Observable<Product[]>;
   constructor(
-    private productService: ProductServiceService
+    private productService: ProductService
   ) { }
 
   ngOnInit(): void {
     this.initializeProducts();
   }
 
+  onListChanged() {
+    this.initializeProducts();
+  }
+
   private initializeProducts() {
-    this.productsInStock = this.productService.getProductsInStock();
+    this.products = this.productService.getProducts();
   }
 
 }