+++ src/app/app.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
+import { HttpClientModule } from '@angular/common/http';
+import { httpInterceptorProviders } from './core/interceptors/index';
 
 import { AppComponent } from './app-component/app.component';
 import { CartModule } from './cart/cart.module';
 import { ProductModule } from './product/product.module';
 import { SharedModule } from './shared/shared.module';
-import { LocalStorageService } from './core/services';
+import { LocalStorageService } from './shared/services';
 import { AboutComponent } from './about/about.component';
 import { OrderModule } from './order/order.module';
 import { AdminModule } from './admin/admin.module';
@@ -21,6 +23,7 @@ import { LoginMinimizedComponent } from './login/components/login-minimized/logi
     LoginMinimizedComponent
   ],
   imports: [
+    HttpClientModule,
     BrowserModule,
     AppRoutingModule,
     CartModule,
@@ -31,7 +34,8 @@ import { LoginMinimizedComponent } from './login/components/login-minimized/logi
   ],
   bootstrap: [AppComponent],
   providers: [
-    { provide: LocalStorageService, useClass: LocalStorageService }
+    { provide: LocalStorageService, useClass: LocalStorageService },
+    httpInterceptorProviders
   ]
 })
 export class AppModule { }