+++ src/app/core/interceptors/index.ts
@@ -1,0 +1,15 @@
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { TimeRequestInterceptor, TimeResponseInterceptor } from './time.interceptor';
+
+export const httpInterceptorProviders = [
+    {
+        provide: HTTP_INTERCEPTORS,
+        useClass: TimeRequestInterceptor,
+        multi: true
+    },
+    {
+        provide: HTTP_INTERCEPTORS,
+        useClass: TimeResponseInterceptor,
+        multi: true
+    }
+];