+++ src/app/shared/interfaces/product.ts
@@ -2,4 +2,5 @@ export interface Product {
     Id: string;
     Name: string;
     Price: number;
+    Remaining: number;
 }