+++ src/app/shared/services/index.ts
@@ -1,4 +1,7 @@
+export * from './guid-helper.service';
 export * from './local-storage.service';
 export * from './config-options.service';
 export * from './constants.service';
 export * from './generator.service';
+export * from './app-settings.service';
+export * from './http-client.service';