+++ src/app/shared/services/local-storage.service.ts
@@ -2,11 +2,13 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class LocalStorageService {
+  // TODO change to storing in constants
   keys = {
     purchasedProducts: 'PurchasedProducts',
     loginInfo: 'LoginInfo',
     productsInDb: 'productsInDb',
-    orderClientData: 'OrderClientData'
+    orderClientData: 'OrderClientData',
+    appSettings: 'appSettings'
   };
 
   setItem(key: string, value: any) {